Added save file scanning and verification to be added to the world list
- Added the functionality of scanning for valid save files in the "saves" folder and displaying them on the list
- Fixed the world listing buttons so that they activate after the world listing is selected
- Added a background to the world selection pane listview that shows when there are no server listings. It contains the server image that was there before and some text telling the user to click the + button to add a new world.

NOTE: To test out the server listing, make a copy of your "world" folder and move a copy of the server.cfg file into it. Then after you opened the launcher once, move that world folder to `C:\Users\(Username)\AppData\Roaming\Nitrox\saves`. Rename it to "save0" and make copies of it, incrementing the last number by 1 for each file