Fixed unintended sit animation for remote players
Remote players should now not play the sitting animation when clicking on an unfinished chair (whether that be with a builder or not).

This also fixes the issue where that player gets a simulation lock on the chair despite not sitting down on it, preventing other players from using it when its built.