Temp fix for deconstructing chair while player is in it
This successfully kicks the player off of the chair they're sitting in when someone else deconstructs the chair.

There are still some issues:
- This doesn't work for the bench for some reason.
- While trying to do a similar fix for the beds (deconstructing while someone is in it), I ran into a ton of issues regarding animations and stuff.
- ALSO for some reason, if the player that was kicked out of the chair doesn't move when the other user starts reconstructing the chair, they will sit back down on a non-existent chair (in the same location they were before) and proceed to be broken.

It is for these reasons that I decided I would try looking into the method of preventing deconstruction of these base pieces instead of the kicking-out method. This still has it's own issues, but I'll try to figure it out